feat: better `fullPage` implementation
