+++ packages/screenshot/package.json
@@ -30,6 +30,8 @@
     "@browserless/goto": "^5.13.2",
     "got": "~9.6.0",
     "is-url-http": "~1.2.0",
+    "jimp": "~0.8.3",
+    "merge-img": "~2.1.3",
     "sharp": "~0.23.0",
     "svg-gradient": "~1.0.2"
   },
